在pro模型的提示下修改, 增添

+++ src/sql/05_train_with_time.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION proc_count_train_sales(v_train_id INT, v_target_time INT)
+CREATE OR REPLACE FUNCTION proc_count_train_sales(v_train_id BIGINT, v_target_time BIGINT)
 RETURNS INT AS $$
 DECLARE
     v_total_sold INT;
